requirement.txt
aiosmtplib==2.0.2
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.1.3
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastapi-mail==1.4.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iso8601==1.1.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.3
passlib==1.7.4
pydantic==2.7.2
pydantic-settings==2.3.3
pydantic_core==2.18.3
Pygments==2.18.0
PyJWT==2.8.0
pypika-tortoise==0.1.6
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
setuptools==70.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tortoise-orm==0.21.3
typer==0.12.3
typing_extensions==4.12.1
ujson==5.10.0
union==0.1.10
urllib3==2.2.1
uvicorn==0.30.1
watchfiles==0.22.0
websockets==12.0
